Сменил название метода save() на savePostDraft(), чтобы логично читалось. В PostService принудительно выставил значение false для поля isPublished, т.к. работаем с черновым вариантом. У User убрал @JoinColumn, а @OneToMany добавил параметр (mappedBy = "user") по полю класса Post. Везде, где надо поправил название метода. Тесты подкручены для проверки на публикацию.

+++ src/main/java/com/corporation/controller/UserController.java
@@ -25,7 +25,7 @@ public class UserController {
     private final UserMapper userMapper;
 
     @GetMapping("/{id}")
-    public ResponseEntity<UserDto> getUserById(@PathVariable("id") int id) {
+    public ResponseEntity<UserDto> getUserById(@PathVariable("id") long id) {
         Optional<User> optionalUser = UserService.findById(id);
 
         return optionalUser
